+++ src/yushkevich/java_2/lesson_1/HomeWork/Main.java
@@ -1,0 +1,6 @@
+package yushkevich.java_2.lesson_1.HomeWork;
+
+public class Main {
+    public static void main(String[] args) {
+    }
+}
