+++ src/yushkevich/java_2/lesson_1/HomeWork/Swim.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1;
+package yushkevich.java_2.lesson_1.HomeWork;
 
 public interface Swim {
