+++ src/yushkevich/java_2/lesson_1/Work/Duck.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1;
+package yushkevich.java_2.lesson_1.Work;
 
 public class Duck extends Animal implements Swim {
     int canSwimDistance;
