+++ src/yushkevich/java_2/lesson_1/Work/Main.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1;
+package yushkevich.java_2.lesson_1.Work;
 
 public class Main {
     public static void main(String[] args) {
