+++ src/yushkevich/java_2/lesson_1/Work/Obstacle.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1;
+package yushkevich.java_2.lesson_1.Work;
 
 public abstract class Obstacle {
     protected final int size;
