+++ src/yushkevich/java_2/lesson_1/Work/Road.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1;
+package yushkevich.java_2.lesson_1.Work;
 
 public class Road extends Obstacle {
