+++ src/yushkevich/java_2/lesson_1/Work/Swim.java
@@ -1,0 +1,6 @@
+package yushkevich.java_2.lesson_1.Work;
+
+public interface Swim {
+
+    void swim(int distance);
+}
