+++ src/yushkevich/java_2/lesson_1/Work/Water.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1;
+package yushkevich.java_2.lesson_1.Work;
 
 public class Water extends Obstacle {
     protected Water(int size) {
